Investigating an issue with Env for Geb
