+++ src/Juvix/Compiler/Backend/Geb/Evaluator.hs
@@ -6,12 +6,12 @@ module Juvix.Compiler.Backend.Geb.Evaluator
 where
 
 import Control.DeepSeq
+import Control.Exception qualified as Exception
 import Juvix.Compiler.Backend.Geb.Data.Context as Context
 import Juvix.Compiler.Backend.Geb.Evaluator.Data
 import Juvix.Compiler.Backend.Geb.Evaluator.Error
 import Juvix.Compiler.Backend.Geb.Evaluator.Options
 import Juvix.Compiler.Backend.Geb.Language
-import Juvix.Compiler.Backend.Geb.Translation.FromGebValue
 import Juvix.Compiler.Backend.Geb.Translation.FromSource as Geb
 import Juvix.Compiler.Backend.Geb.Translation.FromSource.Analysis.TypeChecking as Geb
 
@@ -44,8 +44,8 @@ objNoEvalMsg = "Geb objects cannot be evaluated, only morphisms."
 eval' :: Env -> Morphism -> Either JuvixError GebValue
 eval' env m =
   run . runError $
-    runReader env $
-      mapError (JuvixError @EvalError) $
+    mapError (JuvixError @EvalError) $
+      runReader env $
         eval m
 
 nf' :: Env -> Morphism -> Either JuvixError Morphism
@@ -128,26 +128,8 @@ eval morph = case morph of
   MorphismCase c -> do
     vCaseOn <- eval $ c ^. caseOn
     case vCaseOn of
-      GebValueMorphismLeft leftArg -> do
-        let fun' =
-              MorphismLambda
-                Lambda
-                  { _lambdaBody = c ^. caseLeft,
-                    _lambdaVarType = c ^. caseLeftType,
-                    _lambdaBodyType = c ^. caseCodomainType
-                  }
-        fun <- eval fun'
-        apply' fun leftArg
-      GebValueMorphismRight rightArg -> do
-        let fun' =
-              MorphismLambda
-                Lambda
-                  { _lambdaBody = c ^. caseRight,
-                    _lambdaVarType = c ^. caseRightType,
-                    _lambdaBodyType = c ^. caseCodomainType
-                  }
-        fun <- eval fun'
-        apply' fun rightArg
+      GebValueMorphismLeft leftArg -> apply' (c ^. caseLeft) leftArg
+      GebValueMorphismRight rightArg -> apply' (c ^. caseRight) rightArg
       _ ->
         throw
           EvalError
@@ -171,8 +153,11 @@ apply fun' arg' = do
   fun <- eval fun'
   case fun of
     GebValueClosure cls ->
-      local (over envContext (Context.cons arg)) $
-        eval (cls ^. valueClosureLambdaBody)
+      do
+        let clsEnv = cls ^. valueClosureEnv
+            bodyEnv = Context.cons arg clsEnv
+        local (over envContext (const bodyEnv)) $
+          eval (cls ^. valueClosureLambdaBody)
     _ ->
       throw
         EvalError
@@ -183,13 +168,16 @@ apply fun' arg' = do
 
 apply' ::
   Members '[Reader Env, Error EvalError] r =>
-  GebValue ->
+  Morphism ->
   GebValue ->
   Sem r GebValue
-apply' fun arg =
+apply' fun' arg = do
+  fun <- eval fun'
   case fun of
-    GebValueClosure cls ->
-      local (over envContext (Context.cons arg)) $
+    GebValueClosure cls -> do
+      let clsEnv = cls ^. valueClosureEnv
+          bodyEnv = Context.cons arg clsEnv
+      local (over envContext (const bodyEnv)) $
         eval (cls ^. valueClosureLambdaBody)
     _ ->
       throw $
@@ -267,3 +255,87 @@ valueTrue = GebValueMorphismLeft GebValueMorphismUnit
 
 valueFalse :: GebValue
 valueFalse = GebValueMorphismRight GebValueMorphismUnit
+
+-- TODO: DONT REVIEW BELOW YET
+fromGebValueError :: Text -> Maybe GebValue -> Maybe Object -> a
+fromGebValueError msg val gebExpr =
+  Exception.throw
+    FromGebValueError
+      { _fromGebValueErrorMsg = msg,
+        _fromGebValueErrorGebValue = val,
+        _fromGebValueErrorGebExpression = gebExpr
+      }
+
+-- We cannot quote GebValues to Morphisms without knowing the type
+-- in some cases. For example, we cannot quote a GebValueMorphismLeft.
+-- So we need to provide the Geb object associated to the Geb value.
+
+needObjectInfo :: GebValue -> Bool
+needObjectInfo = \case
+  GebValueMorphismUnit -> False
+  GebValueMorphismInteger {} -> False
+  GebValueClosure {} -> False
+  GebValueMorphismLeft {} -> True
+  GebValueMorphismRight {} -> True
+  GebValueMorphismPair {} -> True
+
+-- | Quote GebValues to Morphisms.
+fromGebValue ::
+  Members '[Reader Env, Error JuvixError] r =>
+  Maybe Object ->
+  GebValue ->
+  Sem r Morphism
+fromGebValue ty = \case
+  GebValueMorphismInteger i -> return $ MorphismInteger i
+  GebValueMorphismUnit -> return MorphismUnit
+  GebValueClosure cls -> case ty of
+    Just (ObjectHom funObj) ->
+      return $
+        MorphismLambda
+          Lambda
+            { _lambdaVarType = funObj ^. homDomain,
+              _lambdaBodyType = funObj ^. homCodomain,
+              _lambdaBody =
+                cls ^. valueClosureLambdaBody
+            }
+    Just _ ->
+      fromGebValueError
+        "Got Helper wrong object. Expected function object (lambda)"
+        Nothing
+        ty
+    Nothing -> fromGebValueError "(closure) Need object info" Nothing ty
+  val@(GebValueMorphismLeft m) -> case ty of
+    Just (ObjectCoproduct _) -> MorphismLeft <$> fromGebValue ty m
+    Just _ ->
+      fromGebValueError
+        "type mismatch (left). Expected a coproduct"
+        (Just val)
+        ty
+    Nothing -> fromGebValueError "need object info" (Just val) ty
+  val@(GebValueMorphismRight m) -> case ty of
+    Just (ObjectCoproduct _) -> MorphismRight <$> fromGebValue ty m
+    Just _ ->
+      fromGebValueError
+        "type mismatch (right). Expected a coproduct"
+        (Just val)
+        ty
+    Nothing -> fromGebValueError "need object info" (Just val) ty
+  val@(GebValueMorphismPair m) -> case ty of
+    Just (ObjectProduct prod) -> do
+      let (a, b) = (prod ^. productLeft, prod ^. productRight)
+      pLeft <- fromGebValue (Just a) (m ^. valueMorphismPairLeft)
+      pRight <- fromGebValue (Just b) (m ^. valueMorphismPairRight)
+      return $
+        MorphismPair
+          Pair
+            { _pairLeft = pLeft,
+              _pairRight = pRight,
+              _pairLeftType = a,
+              _pairRightType = b
+            }
+    Just _ ->
+      fromGebValueError
+        "type mismatch (pair). Expected a product"
+        (Just val)
+        ty
+    Nothing -> fromGebValueError "need object info" (Just val) ty
