+++ src/Juvix/Compiler/Backend/Geb/Pretty/Keywords.hs
@@ -55,6 +55,15 @@ kwPair = keyword Str.gebPair
 kwLamb :: Doc Ann
 kwLamb = keyword Str.gebLamb
 
+kwClosure :: Doc Ann
+kwClosure = keyword Str.gebValueClosure
+
+kwClosureEnv :: Doc Ann
+kwClosureEnv = keyword Str.gebValueClosureEnv
+
+kwNil :: Doc Ann
+kwNil = keyword Str.lispNil
+
 kwApp :: Doc Ann
 kwApp = keyword Str.gebApp
