+++ tests/Geb/positive/app-case-on.geb
@@ -2,21 +2,15 @@
     (coprod int int)
     (coprod so1 int)
     (lamb
-        (coprod int int)
-        (coprod so1 int)
+        int int
         (case-on 
             int
             int 
             int
             (index 0)
-            ( lamb
-                int 
-                (coprod so1 int)
-                (right 1))
-             ( lamb
-                int 
-                (coprod so1 int)
-                (right 2))
-        )
+            (right 1)
+            (left 2)
+            )
     )
-    (left 3))
+  (left
+    3))
