+++ tests/Geb/positive/lamb.geb
@@ -1,1 +1,9 @@
-(lamb so1 (!-> so1 so1) (lamb so1 so1 (index 1)))
+(lamb
+  so1
+  (!->
+    so1
+    so1)
+  (lamb
+    so1
+    so1
+    (index 1)))
