+++ tests/Geb/positive/test003.geb
@@ -1,0 +1,1 @@
+(app int int (lamb int int (case-on so1 int int (index 0) (lamb so1 int (right unit)) (lamb int int (case-on int int int (index 1) (lamb int int (app int int (lamb int int (index 0)) (index 2))) (lamb int int (case-on int int int (index 2) (lamb int int (app int int (app int int (lamb int int (lamb int int (app int int (index 0) (index 1)))) (fst int int (index 3))) (snd int int (index 3)))) (lamb int int (app int int (app int int (app int int (lamb int (!-> int int) (lamb int int (lamb int int (app int int (app int int (index 1) (index 2)) (index 0))))) (fst int int (index 3))) (fst int int (snd int int (index 3)))) (snd int int (snd int int (index 3))))))))))) (right (right (right (pair int int (left unit) (pair int int (lamb int int (lamb int int (case-on so1 so1 int (index 0) (lamb so1 int (right unit)) (lamb so1 int (index 2))))) (right unit)))))))
